requirements.txt
# ACORD Data Extractor API

fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
python-dotenv==1.0.1
pypdf==4.2.0
openai>=1.0.0
flask==3.0.3
pydantic>=2.0.0
httpx>=0.24.0